move SP800108KeyGen
